add auth in create, update, delete in programs and fix route programs

+++ app/Http/Controllers/ProgramController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Program;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Request;
 
 class ProgramController extends Controller
@@ -26,6 +27,8 @@ class ProgramController extends Controller
 
     public function store(Request $request)
     {
+        $user = Auth::user();
+
         $validated = $request->validate([
             'name' => 'required|string|max:255',
             'description' => 'required|string|max:255',
@@ -34,8 +37,12 @@ class ProgramController extends Controller
             'restaurant_id' => 'nullable|numeric',
         ]);
         try {
-            $program = Program::create($validated);
-            return response()->json($program, 201);
+            if($user->isAdmin()){
+                $program = Program::create($validated);
+                return response()->json($program, 201);
+            }else{
+                return response()->json(['message' => 'No tens permisos per crear un programa'], 403);
+            }
         } catch (\Exception $e) {
             return response()->json([
                 'message' => 'No se pudo crear el programa debido a un error inesperado.',
@@ -47,33 +54,46 @@ class ProgramController extends Controller
 
     public function update(Request $request, string $id)
     {
+        $user = Auth::user();
         $programs = Program::find($id);
-        if($programs){
 
-            $validated = $request->validate([
-                'name' => 'sometimes|string|max:255',
-                'description' => 'sometimes|string|max:255',
-                'year' => 'sometimes|numeric',
-                'season' => 'sometimes|numeric',
-                'restaurant_id' => 'sometimes|numeric',
-            ]);
+        if($user->isAdmin()){
+            if($programs){
 
-            $programs->fill($validated);
-            $programs->save();
-            return response()->json($programs);
+                $validated = $request->validate([
+                    'name' => 'sometimes|string|max:255',
+                    'description' => 'sometimes|string|max:255',
+                    'year' => 'sometimes|numeric',
+                    'season' => 'sometimes|numeric',
+                    'restaurant_id' => 'sometimes|numeric',
+                ]);
+
+                $programs->fill($validated);
+                $programs->save();
+                return response()->json($programs);
+            }else{
+                return response()->json(['message' => 'El programa no exiteix a les bases de dades'], 404);
+            }
         }else{
-            return response()->json(['message' => 'El programa no exiteix a les bases de dades'], 404);
+            return response()->json(['message' => 'No tens permisos per modificar un programa'], 403);
         }
+
     }
 
     public function destroy(string $id)
     {
+        $user = Auth::user();
         $programs = Program::find($id);
-        if($programs){
-            $programs->delete();
-            return response()->json($programs);
+
+        if($user->isAdmin()){
+            if($programs){
+                $programs->delete();
+                return response()->json($programs);
+            }else{
+                return response()->json(['message' => 'El programa no exiteix a les bases de dades'], 404);
+            }
         }else{
-            return response()->json(['message' => 'El programa no exiteix a les bases de dades'], 404);
+            return response()->json(['message' => 'No tens permisos per eliminar un programa'], 403);
         }
     }
 }
